Now we use Web2.0, shader version updated
I updated shaders from 'version 100'
to 'version 300 es'.
(This is closer to 'version 330' used in desktop,
but still does not work in desktop, and
the code is more similar)

I used glsl-100-to-300 node library.
It generated all ok, except in vert shader
there was a bunch of 'in' instead of 'out',
so I adjusted manually and now we are on version 330 es!